don’t insert IRC replies into the raft log
We now process the raft log as input, but keep output in memory,
separately. Every server in the network generates the same output.

The previous approach was brittle with regards to failures — the raft
leader was the only point in the network which generated responses. If
it failed between applying the IRC input message and applying the IRC
output message(s), you would have lost messages.