refactor IRC command handling
Each command now has an entry in the commands map, pointing to a
function that handles the command itself, a minimum number of parameters
(so that we can check for that in a central place instead of in each
command handler) and a function that defines the routing between
sessions.

Adding a new command can be done entirely in commands.go.

The API for command handlers is two lines shorter per message than the
previous code.