refactor irc output into the outputstream package
This gets rid of all data races in the IRC output handling (ran
robustirc for a minute under the throughput benchmark, no data races
found in this piece of the code).

Further, it adds a way to delete output messages for IRC input messages
once we compact the input away and thereby fixes #79.

We now also avoid a bottleneck: the map in which messages are stored is
now keyed with an int64, for which the Go runtime has a fast path
(previously we used an entire types.RobustId).

The outputstream package has 100% statement test coverage.