compaction: skip log entries that result in an error
There are two cases here: either the message was not found, or some
other LevelDB-internal error happened, which is either a bug or data
corruption.

In case the message is actually corrupt, we’d probably either skip it or
mark it as a message of death and skip it.

The case where the message was not found is not actually a problem. The
code previously had the assumption that consecutive entries always
exist, which is not the case after the first compaction.