Refactor InterestedIn into InterestedFor, add tests.
Before this commit, InterestedIn was mostly a stub, but the idea was
that we’d look at the current state and decide who gets which message.
This approach is problematic, as it requires clients to always get a
message immediately after it is available. This is what happened in real
life so far, but it does not match our architecture. See also
TestInterestedInDelayed() in ircserver/ircserver_test.go for an
illustration of the problem.

With this commit, we store a map called InterestedFor with every
message. The InterestedFor map contains the sessions which should
receive the message. InterestedFor is computed once in SendMessages
(based on the current state directly after processing the input message)
and stays constant.

Potential for reducing memory usage later: perhaps the Interested
function could return the same map as before in case nothing changed.
Perhaps a sync/pool could be used for that?

fixes #23 by achieving 98.4% statement coverage in ircserver/commands.go