Refactor ping sending to happen in GetMessages, avoid outputstream.
The advantage is that we don’t insert ping messages that we can never
delete again (unbounded memory growth) and we don’t need to find a
message id for them. This is important for #94, and without this commit,
the assumption that message ids are strictly monotonically increasing is
sometimes violated, i.e. whenever a message comes in very briefly before
SendPing is called and the server clocks are slightly out of sync.