Avoid calling NewRobustMessage() when not the leader.
See the comment for rationale. This is half of the fix for #100.

In the common case (stable leadership), this saves a bit of resources
and lock contention. In the case of leadership transferring, this change
might require the client to resend a message, i.e. incurs a roundtrip
(possibly with exponential backoff).