Significantly refactor compaction for better performance
Instead of doing a full scan over all messages and filtering by session
id, we now build up an index of all messages that are relevant to a
certain session and then just scan these messages.

This brings down a typical compaction run on robustirc.net from >1 hour
to about 10 minutes tops.

It also opens the door to more precise compaction: before this commit,
we could only look at messages of the same session, but with this
commit, we can make the relevant* functions look at messages that are
interesting for the session in question.

In a week, we should remove the slow path from the code to reduce
complexity. By then, we should be reasonably sure that it’s unused.

fixes #112