Refactor canarying process
Before this commit, when canarying, you’d start the new robustirc binary
locally with the -canary flag, let it download all processed messages
from the currently running network, and then diff the messages with how
the local binary processes them. A big downside of that approach is that
the compactions were not aligned, i.e. during normal operation, you
might get diffs even though the code changes did actually not produce
diffs. This is obviously very misleading and annoying.

With this commit, we add a new orchestration tool called
robustirc-canary. You point robustirc-canary to a raft directory, which
you can obtain by copying the raft directory of a running network, or by
copying the raft directory of a robustirc-localnet instance.
robustirc-canary then starts the specified old and new RobustIRC
binaries, makes them dump their state (for each message: input, output
including InterestingFor, compaction status) to a JSON file and then
reads in these JSON files and writes a report to stdout.

The JSON-encoded state files are called canary_<hash>.json and are
stored in the specified raft directory. <hash> is the sha256 hash of the
binary which created the file. When running robustirc-canary again,
these files are reused and don’t need to be recreated.

The resulting report contains all messages, but displays only these that
differ in some way:
 • Messages which have a red background are compacted on one side, but
   not the other.
 • Messages which have a pink marker on the right-hand side have
   differing InterestingFor values (the server routes them to different
   sessions).
 • Messages which have a red marker on the right-hand side differ in
   their output text, and maybe also have differing InterestingFor
   values.
You can introspect the InterestingFor values by clicking on the @ symbol
before each output message. If the @ symbol is green, the InterestingFor
values match, if the @ symbol is red, they differ.

An example workflow is:

    go build
    scp -r alp.robustirc.net:/usr/bin/robustirc robustirc.old
    scp -r alp.robustirc.net:/media/persistent/robustirc /tmp/raft
    robustirc-canary -old_executable=./robustirc.old /tmp/raft > /tmp/report.html
    google-chrome /tmp/report.html