outputstream: use LevelDB
The motivation for this change is to reduce RAM usage.

The initial observation is that typically, GetMessages requests don’t
lag far behind, meaning that most reads/writes happen in a small window
of the outputstream messages (say, the last 1 MB).

With that in mind, the outputstream typically consumes hundreds of
megabytes of RAM, and that RAM is entirely cold (never overwritten, very
very rarely read, if at all). While Linux can swap out these pages, not
everyone has swap enabled, and relying on swapping to have a reasonably
sized RAM footprint sounds like somewhat of an antipattern to me.

When using LevelDB, we store the messages in a temporary database. The
database is temporary in the sense that it can be deleted after the
RobustIRC node is stopped, and it will be entirely recreated from
snapshots/raftlog.

While LevelDB uses a cache to speed up access to recently read blocks,
benchmarks&profiling show that caching the _deserialized_ blocks is
worthwhile, otherwise we waste a lot of CPU on deserializing the same
messages over and over again when we have a lot of sessions. Hence, we
cache 1000 deserialized messages.

What LevelDB stores on disk is actually much smaller than what we used
to hold in RAM. The Snappy compression brings down the data value from
hundreds of MB to 70 MB. Using the compact message format
(outputstream.Message), this is further reduced to 38 MB.

This commit effectively reverts the interestingforcache which commit
7784d8ef8 introduced — with our current approach to outputstream, using
the interestingforcache is actually a 1.6x memory increase.

With the throughput test program from
https://github.com/robustirc/benchmark, this commit reduces peak
throughput from ≈ 7500 msgs/s to ≈ 5500 msgs/s.