DeleteRange: use an iterator instead of a for loop
That way, we only delete rows which we know are present. This makes a
huge difference when deleting a large range with many holes. In
practice, we have plenty of holes due to compaction, and deleting a
large range might happen e.g. when all users decide to QUIT at the same
time (i.e. a network goes from, say, 50 long-running sessions to 0
sessions).

AIUI, LevelDB writes tombstones for each entry that you tell it to
delete, regardless of whether the entry actually existed or not. Hence,
using a for loop will write many many unnecessary tombstones.

Before this commit, such an example would take > 6 minutes, during which
no other modifications can be made, i.e. the network would be frozen for
more than clients are typically willing to wait, resulting in what looks
like a netsplit.

After this commit, the same example takes about 1 minute.