Specify a 10s timeout for dialing
net/http uses a 30s timeout, but I think we can be a bit more aggressive
than that — if establishing a connection takes more than 10s, the
connection will not be very useful to run Raft on.

Without this change, RobustIRC networks in our loadtest take well over a
minute to elect a leader (the loadtest does not start with a leader),
which seems excessive.

I can imagine that in real-life situations with packet loss, this change
might also be useful.

related to #81