keep config.SessionExpiration worth of messages, not 1 week
Up until now, we retained 1 week’s worth of messages under the assumption that
we’ll want to implement session resumption for up to 1 week. Always keeping the
messages, even if session resumption wasn’t implemented, seemed like a good idea
to get an impression of feasibility. These results are in: retaining messages
for 1w is feasible for networks with a few hundred users.

We may not get around to implementing session resumption: adoption of RobustIRC
itself has not been great so far, so currently it seems that investing any more
effort into even more advanced features is futile.

But, if we do, we should implement another optimization: the server can ping the
client, and upon receipt of the corresponding pong message discard all messages
up until here, because the client must have read them.

With this commit, messages are kept around for just long enough to rescue a
hanging session in the last second. The SessionExpiration time itself has been
reduced from 30m to 10m, but that will only affect new networks.

To implement session resumption, we can then add a new config setting which
supersedes the current behavior. That setting can then default to 1 week again.

fixes #131