Bugfix: correctly update state on Raft snapshot restore
The ircServer, ircStore and outputStream are re-created during snapshot restore.
Before this commit, only the global variables were updated by the code.
However, the copies in the api.HTTP struct also need to be updated!

Otherwise, the API would panic when accessing the outputStream
(“IRCLog” tab of the status page), and new connections to the network would
consistently fail with the following error message:

  No such RobustIRC session (killed by the network?) (non-recoverable)

(Because session creation still happens successfully via Raft,
but when reading or writing new messages, the API was still working
with the old ircServer instance, which did not get the new session.)

Why did this bug linger for so long? Because it is only triggered when Raft
is unable to stream log entries because the node is so far behind in
replication, and falls back to restoring from snapshot. The replication
threshold is tied to config.TrailingLogs, which defaults to 10k log entries.